fix(stage): clippy — allow too-many-args on the `line` helper + drop a redundant cast
Keeps the stage theme engine (d4ac5d8) clean under `clippy -D warnings`: the
`line(...)` draw helper legitimately takes 8 args (rect + text + colour + align),
and a `u32 -> u32` cast in the scripture NEXT-region width was redundant.

+++ implementation/desktop/crates/selahcue-present/src/stage.rs
@@ -239,6 +239,7 @@ fn fill(f: &mut Frame, x: i32, y: i32, w: u32, h: u32, color: Rgba) {
     });
 }
 
+#[allow(clippy::too_many_arguments)]
 fn line(f: &mut Frame, x: i32, y: i32, w: u32, px: u32, text: &str, color: Rgba, align: TextAlign) {
     if text.is_empty() {
         return;
@@ -463,7 +464,7 @@ fn compose_scripture(
         let nrect = Rect::new(
             nx,
             ny,
-            left_w.saturating_sub((cw + (w as f64 * 0.012) as u32) as u32),
+            left_w.saturating_sub(cw + (w as f64 * 0.012) as u32),
             (h as f64 * 0.16) as u32,
         );
         content_region(frame, next, nrect, TextAlign::Left, theme.muted);
